# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/profile:
    get:
      operationId: getStudentProfile
      tags: [students]
      summary: Get current student's profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateStudentProfile
      tags: [students]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /students/{studentId}:
    get:
      operationId: getStudentById
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students (for companies)
      parameters:
        - name: skills
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"

  # Company Profiles
...
components:
  schemas:
...
    StudentProfile:
      type: object
      required: [id, userId, name, email]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        college:
          type: ["string", "null"]
        degree:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        certifications:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            $ref: "#/components/schemas/ExperienceItem"
        profileCompletion:
          type: integer
        linkedinUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
    StudentProfileInput:
      type: object
      properties:
        college:
          type: string
        degree:
          type: string
        graduationYear:
          type: integer
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        avatarUrl:
          type: string
        resumeUrl:
          type: string
        certifications:
          type: array
          items:
            type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
    Project:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
    ExperienceItem:
      type: object
      required: [company, role]
      properties:
        company:
          type: string
        role:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        current:
          type: boolean
        description:
          type: ["string", "null"]

    # Company
...