# lib/api-zod/src/generated/types/experienceItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExperienceItem {
  company: string;
  role: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  current?: boolean;
  /** @nullable */
  description?: string | null;
}