# lib/api-zod/src/generated/types/studentProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfileInput {
  college?: string;
  degree?: string;
  graduationYear?: number;
  skills?: string[];
  bio?: string;
  avatarUrl?: string;
  resumeUrl?: string;
  certifications?: string[];
  linkedinUrl?: string;
  githubUrl?: string;
}